folders added + Filip's files moved
