+++ styles/App.css
@@ -1,5 +1,5 @@
-@import '../styles/GameModeName.css';
-@import '../styles/MainMenuOptions.css';
+@import '../src/app/uicomponents/gamemodename/GameModeName.css';
+@import '../src/app/uicomponents/mainmenuoptions/MainMenuOptions.css';
 
 body {
     font-family: 'Montserrat', sans-serif;
